==> phyloflows/vignettes/03_diagnostics.Rmd <==
---
title: "phyloflows: Performing MCMC diagnostic checks"
date: "2019-09-10"
author: Xiaoyue Xi and Oliver Ratmann
output: github_document
vignette: >
  %\VignetteIndexEntry{Vignette Title}
  %\VignetteEngine{knitr::rmarkdown}
  %\VignetteEncoding{UTF-8}
---

```{r setup, include = FALSE}
knitr::opts_chunk$set(
  collapse = TRUE,
  comment = "#>"
)
```
```{r, include=TRUE, eval=TRUE, echo=FALSE, warning=FALSE, message=FALSE, tidy=TRUE}
require(knitr)
require(kableExtra)
```
```{r, include=FALSE, eval=FALSE, echo=FALSE}
require(rmarkdown)
setwd('/Users/Oliver/git/phyloscanner/phyloflows/vignettes')
infile <- '03_diagnostics.Rmd'
rmarkdown::render(infile, output_format='pdf_document')
rmarkdown::render(infile, output_format='md_document')
```



This vignette describes how to run a number of diagnostics on **phyloflows**
MCMC output, obtained with the function `phyloflows:::source.attribution.mcmc`.
Please work through the vignette *phyloflows: Estimating transmission flows
under heterogeneous sampling -- a first example* before you go ahead here.

## Getting started

We continue our "First_Example". The following code chunk contains all code
needed, up to running **phyloflows** MCMC routine. The only change is that the
number of iterations is now $50,000$. The MCMC should take about 5 minutes to
run.

```{r, include=TRUE, eval=FALSE, echo=TRUE, tidy=FALSE,message=FALSE, warning=FALSE} 
require(data.table)
require(phyloflows)

data(twoGroupFlows1, package="phyloflows")
dobs <- twoGroupFlows1$dobs
dprior <- twoGroupFlows1$dprior
tmp= copy(dprior)
tmp[,WHO:='REC_SAMPLING_CATEGORY']
dprior[,WHO:='TR_SAMPLING_CATEGORY']
dprior <- rbind(dprior,tmp)
control <- list(seed=42, mcmc.n=5e4, verbose=0)
mc <- phyloflows:::source.attribution.mcmc(dobs, dprior, control)
```
```{r, include=TRUE, eval=TRUE, echo=FALSE, tidy=FALSE, message=FALSE, warning=FALSE}
# save(mc, file='~/git/phyloscanner/phyloflows/data/twoGroupFlows1_mcmc.RData')
data(twoGroupFlows1_mcmc, package="phyloflows")
```

## MCMC: diagnostics
**phyloflow** comes with a function to calculate standard MCMC diagnostics. You
can

1. Make trace plots for all model parameters;
2. or make trace plots for the model parameters with smallest effective sample
   size. This may be useful to avoid generating very large pdf files that you 
   won t be able to open anyway.
3. Make trace plots for values of the log likelihood and log posterior density.
4. Calculate acceptance rates.
5. Remove a burn-in period.
6. Calculate effective sample sizes.
7. Calculate summary statistics (mean, median, quantiles) of the marginal
   posterior densities.
8. Plot marginal posterior densities for the model parameters with smallest
   effective sample size. 
9. Make autocorrelation plots for the model parameters with smallest effective
   sample size. 
   
The syntax is as follows. Look up the help page for the diagnostics function for
a full explanation of the control arguments. 
```{r, include=TRUE, eval=TRUE, echo=TRUE, tidy=FALSE,message=FALSE, warning=FALSE}
outfile.base <- file.path(getwd(),'twoGroupFlows1_mcmc') 
control <- list( burnin.p=0.05, 
                 regex_pars='*', 
				 credibility.interval=0.95, 
				 pdf.plot.all.parameters=TRUE, 
				 pdf.plot.n.worst.case.parameters=1, 
				 pdf.height.per.par=1.2, 
				 outfile.base=outfile.base)
phyloflows:::source.attribution.mcmc.diagnostics(mc=mc, control=control)
```

That's it for now. Use your usual R wizadry to process the output further, and
have a look at the other vignettes.
